Changing the top players to league MVP officially
