+++ nbaspa_app/io/league/schemas.py
@@ -20,3 +20,18 @@ class StandingsOutputSchema(Schema):
     W = fields.Int()
     L = fields.Int()
     W_PCT = fields.Float()
+
+class SummaryQueryArgsSchema(Schema):
+    Season = fields.String(default=CURRENT_SEASON)
+    mode = fields.String(
+        validate=validate.OneOf(["survival", "survival-plus"]), default="survival-plus"
+    )
+    sortBy = fields.String(
+        validate=validate.OneOf(["sum", "mean"])
+    )
+
+class AwardOutputSchema(Schema):
+    PLAYER_ID = fields.Int()
+    IMPACT_mean = fields.Float()
+    IMPACT_sum = fields.Float()
+    RANK = fields.Int()
