+++ nbaspa_app/io/players/routes.py
@@ -5,7 +5,6 @@ from pathlib import Path
 from flask import current_app as app
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-import numpy as np
 import pandas as pd
 
 from nbaspa.data.endpoints import AllPlayers, PlayerInfo, PlayerGameLog
@@ -175,42 +174,6 @@ class CommonPlayerInfo(MethodView):
         loader.load()
 
         return loader.get_data("CommonPlayerInfo").to_dict(orient="records")[0]
-
-@io_players.route("/top")
-class TopPlayers(MethodView):
-    """Get the top players for a given season."""
-
-    @io_players.arguments(sc.SummaryQueryArgsSchema, location="query")
-    @io_players.response(200, sc.TopPlayersOutput(many=True))
-    @io_players.paginate()
-    def get(self, args, pagination_parameters):
-        """Get the top players for a given season."""
-        if args["mode"] == "survival":
-            fpath = Path(
-                app.config["DATA_DIR"], args["Season"], "impact-summary.csv"
-            )
-        elif args["mode"] == "survival-plus":
-            fpath = Path(
-                app.config["DATA_DIR"], args["Season"], "impact-plus-summary.csv"
-            )
-        
-        gameratings = pd.read_csv(fpath, sep="|", index_col=0)
-        gameratings.dropna(inplace=True)
-        gameratings["IMPACT_sum"] = gameratings["IMPACT_sum"].round(3)
-        gameratings["IMPACT_mean"] = gameratings["IMPACT_mean"].round(3)
-
-        if args["sortBy"] == "mean":
-            gameratings.sort_values(by="IMPACT_mean", ascending=False, inplace=True)
-        elif args["sortBy"] == "sum":
-            gameratings.sort_values(by="IMPACT_sum", ascending=False, inplace=True)
-        gameratings["RANK"] = np.arange(1, gameratings.shape[0] + 1)
-
-        output = gameratings.to_dict(orient="records")
-        pagination_parameters.item_count = len(output)
-        
-        return output[
-            pagination_parameters.first_item:(pagination_parameters.last_item + 1)
-        ]
 
 @io_players.route("/gamelog")
 class Gamelog(MethodView):
