+++ nbaspa_app/players/templates/season_home.html
@@ -18,7 +18,7 @@
                 </p>
                 <ul class="menu-list">
                     <li>
-                        <a href="{{ url_for('players_bp.top_players', season=season, page=1) }}">
+                        <a href="{{ url_for('players_bp.mvp', season=season, page=1) }}">
                             MVP
                         </a>
                     </li>
